git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@122965 4525493e-7705-40b1-a816-d608a930855b
Former-commit-id: ea1921c03d01de252923ecc913a264155f13cc1f

+++ Online/Online/Gaucho/src/MonCounter.cpp
@@ -403,7 +403,7 @@ void *MonCounter::de_serialize(void *ptr, char *nam)
     {
       void *pp = AddPtr(p,p->dataoff);
       int bsiz = p->reclen - p->dataoff;
-      h->nel = (bsiz)/sizeof(long);
+      h->nel = (bsiz)/sizeof(long long);
       h->ptr = malloc(bsiz);
       memcpy(h->ptr,pp,bsiz);
       break;
